requirements.txt
streamlit
google-generativeai
kerykeion
geopy
pytz
fpdf
timezonefinder